+++ src/services/api.js
@@ -1,88 +1,7 @@
-import axios from 'axios';
+import axios from "axios";
 
-const baseURL = "http://0.0.0.0:8000/"
-
-const axiosInstance = axios.create({
-    baseURL: baseURL,
-    timeout: 5000,
-	headers: {
-		Authorization: localStorage.getItem('access_token')
-			? 'JWT ' + localStorage.getItem('access_token')
-			: null,
-		'Content-Type': 'application/json',
-		accept: 'application/json',
-	}, 
-
+const api = axios.create({
+  baseURL: "https://cheeryup-api.herokuapp.com/",
 });
 
-axiosInstance.interceptors.response.use(
-	(response) => {
-		return response;
-	},
-	async function (error) {
-		const originalRequest = error.config;
-
-		if (typeof error.response === 'undefined') {
-			alert(
-				'A server/network error occurred. ' +
-					'Looks like CORS might be the problem. ' +
-					'Sorry about this - we will get it fixed shortly.'
-			);
-			return Promise.reject(error);
-		}
-
-		if (
-			error.response.status === 401 &&
-			originalRequest.url === baseURL + 'token/refresh/'
-		) {
-			window.location.href = '/Login/';
-			return Promise.reject(error);
-		}
-
-		if (
-			error.response.data.code === 'token_not_valid' &&
-			error.response.status === 401 &&
-			error.response.statusText === 'Unauthorized'
-		) {
-			const refreshToken = localStorage.getItem('refresh_token');
-
-			if (refreshToken) {
-				const tokenParts = JSON.parse(atob(refreshToken.split('.')[1]));
-
-				// exp date in token is expressed in seconds, while now() returns milliseconds:
-				const now = Math.ceil(Date.now() / 1000);
-				console.log(tokenParts.exp);
-
-				if (tokenParts.exp > now) {
-					return axiosInstance
-						.post('/token/refresh/', { refresh: refreshToken })
-						.then((response) => {
-							localStorage.setItem('access_token', response.data.access);
-							localStorage.setItem('refresh_token', response.data.refresh);
-
-							axiosInstance.defaults.headers['Authorization'] =
-								'JWT ' + response.data.access;
-							originalRequest.headers['Authorization'] =
-								'JWT ' + response.data.access;
-
-							return axiosInstance(originalRequest);
-						})
-						.catch((err) => {
-							console.log(err);
-						});
-				} else {
-					console.log('Refresh token is expired', tokenParts.exp, now);
-					window.location.href = '/Login/';
-				}
-			} else {
-				console.log('Refresh token not available.');
-				window.location.href = '/Login/';
-			}
-		}
-
-		// specific error handling done elsewhere
-		return Promise.reject(error);
-	}
-);
-
-export default axiosInstance;
+export default api;
